File Management + Adding Recoil
The entire Board.tsx had all the functions and made the code very big ... this commit aims to breakdown board.tsx to many other smaller helper components. Each file in the subdirectory helper contains a small portion of Board.tsx. Allthough all states and state update function has to be passed in function which increase the number of props ... this needs to be handled somehow . Will be taken care of in future commits i guess :)